Code structure overhaul:
media.py has become too unwieldy. Individual modules will be created for track, media, video, playlist
